Relocate tab props
