+++ src/components/editor/Bindings/Editor.tsx
@@ -15,8 +15,7 @@ import ResourceConfig from 'components/collection/ResourceConfig';
 import MessageWithLink from 'components/content/MessageWithLink';
 import DiscoveredSchemaCommands from 'components/editor/Bindings/SchemaEditCommands/DiscoveredSchema';
 import ExistingSchemaCommands from 'components/editor/Bindings/SchemaEditCommands/ExistingSchema';
-import BindingsTabs from 'components/editor/Bindings/Tabs';
-import { tabProps } from 'components/editor/Bindings/types';
+import BindingsTabs, { tabProps } from 'components/editor/Bindings/Tabs';
 import { useEditorStore_persistedDraftId } from 'components/editor/Store/hooks';
 import AlertBox from 'components/shared/AlertBox';
 import ButtonWithPopper from 'components/shared/ButtonWithPopper';
