+++ src/components/editor/Bindings/types.ts
@@ -1,16 +1,6 @@
 export type TabOptions = 'config' | 'schema';
 
-interface MuiTabProps {
+export interface MuiTabProps {
     label: string;
     value: TabOptions;
 }
-export const tabProps: MuiTabProps[] = [
-    {
-        label: 'workflows.collectionSelector.tab.resourceConfig',
-        value: 'config',
-    },
-    {
-        label: 'workflows.collectionSelector.tab.collectionSchema',
-        value: 'schema',
-    },
-];
